chore: add brewfile (#16)
This pull request primarily focuses on improving the setup process for
OSX users by leveraging Homebrew. The major changes include the addition
of a `Brewfile` which specifies the necessary dependencies, and an
update to the `README.md` file to guide OSX users on how to install
these prerequisites.

*
[`Brewfile`](diffhunk://#diff-1b33d9c13a9a1b0c5d5bc83697ba7b4d36e6cf45f78184c3a28527ffc4418e2dR1-R4):
Added a new file that contains a list of necessary dependencies for the
project, including `kubectl`, `kustomize`, and `helm`. This allows OSX
users with Homebrew installed to easily setup the required tools.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R19-R20):
Updated the prerequisites section to include instructions for OSX users
on how to install the prerequisites using the `brew bundle` command.
This simplifies the setup process for these users.