fix: update OTEL Operator CRD to v1beta1 (#28)
This pull request includes several updates to the OpenTelemetry
Collector configurations and Kustomize files to support the latest
versions and improve the configuration structure. The most important
changes are grouped into version updates and configuration adjustments.

### Version Updates:
* Updated the OpenTelemetry Operator version from `v0.98.0` to
`v0.107.0` in `cluster-infra/otel-operator/kustomization.yaml`. This
version supports `v1beta1`.
* Updated the `apiVersion` from `opentelemetry.io/v1alpha1` to
`opentelemetry.io/v1beta1` and the collector image version from
`0.58.0-amd64` to `0.66.2-amd64` in `collectors/base/collector.yaml`.

### Configuration Adjustments:
Currently ConfigMapGenerator set the spec.config of the collectors as a
multi-line string. `v1beta1` looks for config objects instead of string
which throws errors previously. Instead of generating a config map which
each `colconfig.yaml` for the collectors. We patch the configuration
with kustomize to generate the proper configuration `v1beta1` of the
OTEL CRD is looking for.

CRD Changelog:
https://github.com/open-telemetry/opentelemetry-operator/blob/main/docs/crd-changelog.md